Complete E21 E22

+++ product-service/src/service/ProductService.ts
@@ -6,6 +6,7 @@ import { AppValidationError } from '../utility/errors'
 import { ProductInput } from '../dto/ProductInput'
 import { ServiceInput } from '../dto/ServiceInput'
 import { CategoryRepository } from '../repository/CategoryRepository'
+import { AuthUser } from '../utility/auth'
 
 export class ProductService {
   _repository: ProductRepository
@@ -17,12 +18,28 @@ export class ProductService {
     return ErrorResponse(404, new Error('method not allowed!'))
   }
 
+  async authorisedUser(user_id: number, productId: string) {
+    const product = await this._repository.getProductById(productId)
+    if (!product) return false
+    return Number(user_id) === Number(product.seller_id)
+  }
+
   async createProduct(event: APIGatewayEvent) {
-    const input = plainToClass(ProductInput, JSON.parse(event.body!))
+    // Validate input
+    const token = event.headers.Authorization
+    const user = await AuthUser(token)
+    if (!user) return ErrorResponse(403, 'Please login first')
+    if (user.user_type.toUpperCase() !== 'SELLER')
+      return ErrorResponse(403, 'Only seller can create product')
+
+    const input = plainToClass(ProductInput, event.body)
     const error = await AppValidationError(input)
     if (error) return ErrorResponse(404, error)
 
-    const data = await this._repository.createProduct(input)
+    const data = await this._repository.createProduct({
+      ...input,
+      seller_id: user.user_id,
+    })
 
     await new CategoryRepository().addItem({
       id: input.category_id,
@@ -36,6 +53,17 @@ export class ProductService {
     return SuccessResponse(data)
   }
 
+  async getSellerProducts(event: APIGatewayEvent) {
+    const token = event.headers.Authorization
+    const user = await AuthUser(token)
+    if (!user) return ErrorResponse(403, 'Please login first')
+    if (user.user_type.toUpperCase() !== 'SELLER')
+      return ErrorResponse(403, 'You are not authorised to view this page')
+
+    const data = await this._repository.getAllSellerProducts(user.user_id)
+    return SuccessResponse(data)
+  }
+
   async getProduct(event: APIGatewayEvent) {
     const productId = event.pathParameters?.id
     if (!productId) return ErrorResponse(403, 'please provide product id')
@@ -45,12 +73,23 @@ export class ProductService {
   }
 
   async editProduct(event: APIGatewayEvent) {
+    // Validate input
+    const token = event.headers.Authorization
+    const user = await AuthUser(token)
+    if (!user) return ErrorResponse(403, 'Please login first')
+    if (user.user_type.toUpperCase() !== 'SELLER')
+      return ErrorResponse(403, 'Only seller can edit product')
+
     const productId = event.pathParameters?.id
     if (!productId) return ErrorResponse(403, 'please provide product id')
 
-    const input = plainToClass(ProductInput, JSON.parse(event.body!))
+    const input = plainToClass(ProductInput, event.body)
     const error = await AppValidationError(input)
     if (error) return ErrorResponse(404, error)
+
+    const isAuthorised = await this.authorisedUser(user.user_id, productId)
+    if (!isAuthorised)
+      return ErrorResponse(403, 'You are not authorised to edit this product')
 
     input.id = productId
     const data = await this._repository.updateProduct(input)
@@ -59,13 +98,24 @@ export class ProductService {
   }
 
   async deleteProduct(event: APIGatewayEvent) {
+    // Validate input
+    const token = event.headers.Authorization
+    const user = await AuthUser(token)
+    if (!user) return ErrorResponse(403, 'Please login first')
+    if (user.user_type.toUpperCase() !== 'SELLER')
+      return ErrorResponse(403, 'Only seller can delete product')
+
     const productId = event.pathParameters?.id
     if (!productId) return ErrorResponse(403, 'please provide product id')
+
+    const isAuthorised = await this.authorisedUser(user.user_id, productId)
+    if (!isAuthorised)
+      return ErrorResponse(403, 'You are not authorised to delete this product')
 
     const { category_id, deleteResult } = await this._repository.deleteProduct(
       productId,
     )
-    await new CategoryRepository().addItem({
+    await new CategoryRepository().removeItem({
       id: category_id,
       products: [productId],
     })
